requirements.txt
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=0.19.0